Move TypedData to common.typeddata
